# openapi.yaml
openapi: 3.0.3
info:
  title: Doctor.mx API
  description: |
    Doctor.mx - Plataforma integral de telemedicina para México.
    
    Esta API proporciona acceso a todas las funcionalidades de la plataforma Doctor.mx,
    incluyendo consultas médicas virtuales, gestión de citas, recetas digitales,
    análisis de IA, consentimientos digitales, y más.
    
    ## Autenticación
    La API utiliza Supabase Auth para la autenticación. Incluye el token de sesión
    en el header `Authorization: Bearer <token>`.
    
    ## Roles
    - **patient**: Pacientes que buscan atención médica
    - **doctor**: Médicos que brindan consultas
    - **admin**: Administradores del sistema
    
    ## Rate Limiting
    Las solicitudes están limitadas por IP y usuario autenticado.
    Headers de respuesta:
    - `X-RateLimit-Limit`: Límite de solicitudes
    - `X-RateLimit-Remaining`: Solicitudes restantes
    - `X-RateLimit-Reset`: Tiempo de reinicio
  version: 1.0.0
  contact:
    name: Doctor.mx Support
    email: soporte@doctor.mx
  license:
    name: Propietario
    url: https://doctor.mx/terms
servers:
  - url: https://doctor.mx
    description: Producción
  - url: https://staging.doctor.mx
    description: Staging
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/consent/history:
    get:
      summary: Historial de consentimientos
      description: Obtiene el historial de cambios de consentimientos
      tags:
        - Consent
      responses:
        '200':
          description: Historial
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsentHistory'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/consent/version/{type}:
    get:
      summary: Versión de consentimiento
      description: Obtiene la versión actual de un tipo de consentimiento
      tags:
        - Consent
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Versión de consentimiento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentVersion'

  # ==================== ANALYTICS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de sesión de Supabase
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Cookie de sesión de Supabase
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-01-15T10:30:00Z"
...
    ConsentType:
      type: string
      enum: [medical_treatment, data_processing, telemedicine, recording, ai_analysis, data_sharing, research, marketing, emergency_contact, prescription_forwarding]
...
    ConsentHistory:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        consent_id:
          $ref: '#/components/schemas/UUID'
        action:
          type: string
          enum: [granted, withdrawn, updated]
        performed_at:
          $ref: '#/components/schemas/Timestamp'
        ip_address:
          type: string
        user_agent:
          type: string
    ConsentVersion:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        consent_type:
          $ref: '#/components/schemas/ConsentType'
        version:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        effective_date:
          $ref: '#/components/schemas/Timestamp'
        required_for_new_users:
          type: boolean

    # ==================== REVIEWS ====================
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
...
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Se requiere autenticación"
...